Defunct functions in searcher

Functions listed below are no longer included in the searcher package

Details

search_ixquick(): The function binding was removed as the search engine name changed to "Startpage". Please use search_startpage().

search_rstudio_community(): The function binding was removed as the company name changed to "Posit". Please use search_posit_community().

search_rscom(): The function binding was removed as the company name changed to "Posit". Please use search_posit().